+++ backend/src/main/kotlin/com/puzzle/backend/oauth/controller/UsersControllerSpec.kt
@@ -1,7 +1,7 @@
-package com.puzzle.backend.common.oauth.controller
+package com.puzzle.backend.oauth.controller
 
 import com.puzzle.backend.common.BaseResponse
-import com.puzzle.backend.common.oauth.dto.response.LoginSuccessResponse
+import com.puzzle.backend.oauth.dto.response.LoginSuccessResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.tags.Tag
