+++ backend/src/main/kotlin/com/puzzle/backend/oauth/service/UsersService.kt
@@ -1,8 +1,8 @@
-package com.puzzle.backend.common.oauth.service
+package com.puzzle.backend.oauth.service
 
 import com.puzzle.backend.common.BaseResponse
-import com.puzzle.backend.common.oauth.dto.response.LoginSuccessResponse
-import com.puzzle.backend.common.oauth.dto.response.RefreshDataResponse
+import com.puzzle.backend.oauth.dto.response.LoginSuccessResponse
+import com.puzzle.backend.oauth.dto.response.RefreshDataResponse
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
